V1.4.0_151107 本次修订版主要解决以下问题： 1.新增WSTMall商品详情商品二维码。 2.新增店铺、商品的收藏功能。 3.新增在线支付的流水号号记录。 4.详情新增批量上传图片的功能。 5.新增商品导入功能。 6.后台商品列表新增“精品”“推荐”商品筛选,新增系统分类的筛选。 7.新增商品批量操作订单的操作功能，如-受理、打包、配送的订单操作。 8.购物车中加入店铺QQ。 9.调整注册页、登录页布局。 10.期望送达时间与店铺的营业时间关联。 11.修复购物车两个全选功没起作用的bug。 12.修复商品详情页图片没限制宽度体验性欠佳的BUG。 13.修复搜索时输入空格被转为+号，导致搜索结果不正确的BUG。 14.删除店铺的时候一起删除店铺的登录账号并且下架店铺的所有商品。 15.把店铺改为未审核或者拒绝时，也将店铺的商品状态改为下架以及未审核。 16.商品搜索时搜索关键字可通过空格分割进行搜索。 17.修复购物车修改购买数量时总价格错误的BUG。 18.我是卖家-店铺设置-滚动广告，给图片增加点击的链接。 19.修复商品详情，商品介绍没有设置自动换行的bug。 20.商品详情-商品评价排序修改。 21.修复注册无法输入@的bug。 22.修复已删除店铺在店铺街还显示的bug。 23.修复用户提到的首页图片模糊问题。 24.商品名称输入长度增加长度。 25.修复商家二级商品分类双击隐藏失败的bug。 26.修复商品搜索获取价格分类出错的bug。 27.修复在线支付失败的bug。 28.后台商品列表修改。
*若已安装的用户请参照以下步骤：
1.执行wstmall/Upgrade/1.4.0.sql进行数据库升级。
2.将wstmall/Upgrade/Upgrade里面的程序拷贝到网站WSTMall根目录下覆盖。

+++ Apps/Admin/Model/AdsModel.class.php
@@ -14,9 +14,9 @@ class AdsModel extends BaseModel {
 	  */
 	 public function insert(){
 	 	$rd = array('status'=>-1);
-	 	$id = I("id",0);
+	 	$id = (int)I("id",0);
 		$data = array();
-		$data["adPositionId"] = I("adPositionId");
+		$data["adPositionId"] = (int)I("adPositionId");
 		$data["adFile"] = I("adFile");
 		$data["adStartDate"] = I("adStartDate");
 		$data["adEndDate"] = I("adEndDate");
@@ -39,19 +39,19 @@ class AdsModel extends BaseModel {
 	  */
 	 public function edit(){
 	 	$rd = array('status'=>-1);
-	 	$id = I("id",0);
-		$data["adPositionId"] = I("adPositionId");
+	 	$id = (int)I("id",0);
+		$data["adPositionId"] = (int)I("adPositionId");
 		$data["adFile"] = I("adFile");
 		$data["adStartDate"] = I("adStartDate");
 		$data["adEndDate"] = I("adEndDate");
-		$data["adSort"] = I("adSort",0);
+		$data["adSort"] = (int)I("adSort",0);
 	    if($this->checkEmpty($data,true)){	
 	    	$data["adName"] = I("adName");
 			$data["adURL"] = I("adURL");
-	    	$data["areaId1"] = I("areaId1");
-			$data["areaId2"] = I("areaId2");
+	    	$data["areaId1"] = (int)I("areaId1");
+			$data["areaId2"] = (int)I("areaId2");
 			$m = M('ads');
-		    $rs = $m->where("adId=".I('id',0))->save($data);
+		    $rs = $m->where("adId=".(int)I('id',0))->save($data);
 			if(false !== $rs){
 				$rd['status']= 1;
 			}
@@ -63,13 +63,13 @@ class AdsModel extends BaseModel {
 	  */
      public function get(){
 	 	$m = M('ads');
-		return $m->where("adId=".I('id'))->find();
+		return $m->where("adId=".(int)I('id'))->find();
 	 }
 	 /**
 	  * 分页列表
 	  */
      public function queryByPage(){
-     	$adPositionId = I('adPositionId');
+     	$adPositionId = (int)I('adPositionId');
      	$adDateRange = I('adDateRange');
      	$adName = I('adName');
         $m = M('ads');
@@ -99,7 +99,7 @@ class AdsModel extends BaseModel {
 	 public function del(){
 	    $rd = array('status'=>-1);
 	    $m = M('ads');
-	    $rs = $m->delete(I('id'));
+	    $rs = $m->delete((int)I('id'));
 		if(false !== $rs){
 		   $rd['status']= 1;
 		}
